feat(python): rename PyPI package from webflow_webflow to pulumi_webflow
The Python SDK package name was using the default Pulumi codegen naming
convention (namespace_name) which resulted in the awkward "webflow_webflow"
package name. This change adds explicit Python language configuration to
use "pulumi_webflow" instead, following the standard Pulumi provider
naming convention (e.g., pulumi_aws, pulumi_gcp).

Changes:
- Add Python language configuration to provider.go with packageName
- Regenerate all SDKs via make codegen
- Rename sdk/python/webflow_webflow/ to sdk/python/pulumi_webflow/
- Update setup.py with new package name
